refactor: modularize dev menu
