+++ lib/screens/dev_menu/dev_menu_section.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/widgets.dart';
+
+class DevMenuSection {
+  final String title;
+  final WidgetBuilder builder;
+
+  const DevMenuSection({required this.title, required this.builder});
+}
